+++ packages/core-cairo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.13.0 (2024-05-22)
+
+- **Breaking changes**:
+  - Use OpenZeppelin Contracts for Cairo v0.13.0. ([#359](https://github.com/OpenZeppelin/contracts-wizard/pull/359))
+  - Use Hooks with ERC721 and ERC1155.
+  - Use Hooks for Pausable.
+
 ## 0.12.0 (2024-05-01)
 
 - Add `votes` option to ERC20. ([#355](https://github.com/OpenZeppelin/contracts-wizard/pull/355))